Control.Distributed.Fork.AWS

Usage

withLambdaBackend :: LambdaBackendOptions -> (Backend -> IO a) -> IO a

Provides a Backend using AWS Lambda functions.

Options

data LambdaBackendOptions

Options required for creating a Lambda backend.

lambdaBackendOptions

Arguments

:: Text
Name of the S3 bucket to store the deployment archive in.

-> LambdaBackendOptions

lboPrefix :: Lens' LambdaBackendOptions Text

Prefix to the deployment archive and the CloudFormation stack.

lboMemory :: Lens' LambdaBackendOptions Int

Desired memory for the Lambda functions.

lboMaxConcurrentInvocations :: Lens' LambdaBackendOptions Int

Maximum number of concurrent "invoke" calls to AWS API to trigger executions.

lboMaxConcurrentExecutions :: Lens' LambdaBackendOptions Int

Maximum number of concurrently executing Lambda functions.

lboMaxConcurrentDownloads :: Lens' LambdaBackendOptions Int

If the size of the return value from your function is larger than 200 kilobytes, we fetch the results via S3. This parameter sets the maximum number of concurrent downloads from S3.

lboKeepStack :: Lens' LambdaBackendOptions Bool

Whether to keep the CloudFormation stack after the withLambdaBackend call. Useful for debugging.